requirements.txt
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.27.0
transformers>=4.40.0
accelerate>=0.30.0
xformers>=0.0.24
safetensors>=0.4.0
huggingface_hub>=0.22.0

# Data processing
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
OpenEXR>=1.3.9
Imath>=2.5.7

# Preprocessing tools
sam2>=0.4.0

# Demo
gradio>=4.20.0

# Misc
tqdm>=4.66.0
tensorboard>=2.14.0
omegaconf>=2.3.0
einops>=0.7.0
